feat: refactored nav item
