+++ src/components/Nav/Nav.tsx
@@ -7,20 +7,11 @@ import Logo from '@icons/Logo'
 import Share from '@icons/Share'
 import { useState } from 'react'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
+import NavItem from './components/NavItem'
 import * as S from './Nav.style'
 
 function Nav() {
   const { isSmall } = useScreenContext()
-
-  const projectList = useRecoilValue(projectAtom)
-  const skills = useRecoilValue(skillsAtom)
-
-  const setDimdModal = useSetRecoilState(modalAtom)
-
-  const copyURLtoClipboard = async () => {
-    await navigator.clipboard.writeText(window.location.href)
-    setDimdModal((prev) => ({ ...prev, dimd: true }))
-  }
 
   if (isSmall)
     return (
@@ -36,17 +27,7 @@ function Nav() {
       <ConfirmModal alertText="Copied to clipboard!" />
       <S.NavWrapper>
         <Logo />
-        <S.NavList>
-          <S.NavItem>
-            <span>Skills</span>
-            <S.Badge>{skills.length}</S.Badge>
-          </S.NavItem>
-          <S.NavItem>
-            <span>Projects</span>
-            <S.Badge>{projectList.length}</S.Badge>
-          </S.NavItem>
-          <Share onClick={copyURLtoClipboard} />
-        </S.NavList>
+        <NavItem />
       </S.NavWrapper>
     </S.NavContainer>
   )
